+++ OnlineUIComponents/ArenaSpawnLocationIndicator.cs
@@ -19,11 +19,12 @@ public class ArenaSpawnLocationIndicator : HudPart
     public override void Update()
     {
         counter++;
-        if (counter != 30) return;
+        if (counter != 10) return;
 
         if (clientSettings.avatars.Count == 0
             || clientSettings.avatars[0]?.FindEntity(true) is not OnlineCreature oc
-            || oc.realizedCreature is not Player player)
+            || oc.realizedCreature is not Player player
+            || player.inShortcutVessel is not null)
         {
             counter = 0;
             return;